[ML] Transforms: Fix pivot preview table mapping. (#60609)
- Fixes regression caused by elastic/elasticsearch#53572.
- Adjusts the TS mappings and code to reflect the newly returned API response.
- Re-enables functional tests.
