+++ x-pack/plugins/transform/public/app/common/index.ts
@@ -36,6 +36,7 @@ export {
   TRANSFORM_MODE,
 } from './transform_stats';
 export { getDiscoverUrl } from './navigation';
+export { GetTransformsResponse, PreviewData, PreviewMappings } from './pivot_preview';
 export {
   getEsAggFromAggConfig,
   isPivotAggsConfigWithUiSupport,
