+++ x-pack/plugins/transform/public/app/components/pivot_preview/use_pivot_preview_data.ts
@@ -9,8 +9,7 @@ import { useEffect, useState } from 'react';
 import { dictionaryToArray } from '../../../../common/types/common';
 import { useApi } from '../../hooks/use_api';
 
-import { Dictionary } from '../../../../common/types/common';
-import { IndexPattern, ES_FIELD_TYPES } from '../../../../../../../src/plugins/data/public';
+import { IndexPattern } from '../../../../../../../src/plugins/data/public';
 
 import {
   getPreviewRequestBody,
@@ -18,6 +17,8 @@ import {
   PivotAggsConfigDict,
   PivotGroupByConfigDict,
   PivotQuery,
+  PreviewData,
+  PreviewMappings,
 } from '../../common';
 
 export enum PIVOT_PREVIEW_STATUS {
@@ -27,27 +28,12 @@ export enum PIVOT_PREVIEW_STATUS {
   ERROR,
 }
 
-interface EsMappingType {
-  type: ES_FIELD_TYPES;
-}
-
-export type PreviewItem = Dictionary<any>;
-type PreviewData = PreviewItem[];
-interface PreviewMappings {
-  properties: Dictionary<EsMappingType>;
-}
-
 export interface UsePivotPreviewDataReturnType {
   errorMessage: string;
   status: PIVOT_PREVIEW_STATUS;
   previewData: PreviewData;
   previewMappings: PreviewMappings;
   previewRequest: PreviewRequestBody;
-}
-
-export interface GetTransformsResponse {
-  preview: PreviewData;
-  mappings: PreviewMappings;
 }
 
 export const usePivotPreviewData = (
@@ -77,9 +63,9 @@ export const usePivotPreviewData = (
     setStatus(PIVOT_PREVIEW_STATUS.LOADING);
 
     try {
-      const resp: GetTransformsResponse = await api.getTransformsPreview(previewRequest);
+      const resp = await api.getTransformsPreview(previewRequest);
       setPreviewData(resp.preview);
-      setPreviewMappings(resp.mappings);
+      setPreviewMappings(resp.generated_dest_index.mappings);
       setStatus(PIVOT_PREVIEW_STATUS.LOADED);
     } catch (e) {
       setErrorMessage(JSON.stringify(e, null, 2));
