+++ x-pack/test/functional/apps/transform/index.ts
@@ -8,8 +8,7 @@ import { FtrProviderContext } from '../../ftr_provider_context';
 export default function({ getService, loadTestFile }: FtrProviderContext) {
   const transform = getService('transform');
 
-  // prevent test failures with current ES snapshot, see https://github.com/elastic/kibana/issues/60516
-  describe.skip('transform', function() {
+  describe('transform', function() {
     this.tags(['ciGroup9', 'transform']);
 
     before(async () => {
